// src/enums/SecureStoreEntry.ts
enum SecureStoreEntry {
  BASE_API_URL = "BASE_API_URL",  
  API_KEY = "API_KEY",
  THEME = "THEME",
  LOGS_REFRESH_INTERVAL = "LOGS_REFRESH_INTERVAL",
  LOGS_SINCE = "LOGS_SINCE",
  LOGS_MAX_LINES = "LOGS_MAX_LINES",
  CONTAINER_ORDER_BY = "CONTAINER_ORDER_BY",
  STACK_ORDER_BY = "STACK_ORDER_BY",
  SELECTED_ENDPOINT_ID = "SELECTED_ENDPOINT_ID"
}
  
export default SecureStoreEntry;